[feat] Support delete functionality

+++ app/assets/javascripts/components/user_list.es6.jsx
@@ -1,14 +1,30 @@
 class UserList extends React.Component {
   constructor(props) {
     super(props)
+    this.state = { ... props }
+  }
+
+  handleDelete(id) {
+    console.log("Delete " + id);
+    var users = this.state.users.filter((user) => {
+      return !(user.id === id);
+    })
+    this.setState({
+      users: users
+    })
   }
 
   render () {
   	users = []
-  	for (var i=0; i < this.props.users.length; i++) {
-  		var user = this.props.users[i]
+  	for (var i=0; i < this.state.users.length; i++) {
+  		var user = this.state.users[i]
   		users.push(
-  			<ListRow key={user.id} user={user} current={this.props.user.id} />
+  			<ListRow 
+          key={user.id} 
+          user={user} 
+          current={this.state.user.id} 
+          handleDelete={(id) => this.handleDelete(id)}
+          />
   		);
   	}
     return (
@@ -33,20 +49,26 @@ class ListRow extends React.Component {
 
   constructor(props) {
     super(props)
-    var edit = "/users/" + this.props.user.id + "/edit"
-    var del = "/users/" + this.props.user.id
+    var edit = "users/" + this.props.user.id + "/edit"
+    var del = "users/" + this.props.user.id
     this.state = {
       edit: edit,
       delete: del
     }
-    console.log(this.state)
   }
 
   handleDelete() {
-    console.log("Deleting " + this.state.delete)
-    fetch(this.state.delete, {
-      method: 'DELETE'
-    })
+    var token = document.getElementsByName("csrf-token")[0].content;
+    if (this.state.delete != null) {
+      fetch(this.state.delete, {
+        method: 'DELETE',
+        headers: {
+          'X-CSRF-TOKEN': token
+        },
+        credentials: 'same-origin'
+      })
+    }
+    this.props.handleDelete(this.props.user.id)
   }
 
 	render () {
@@ -54,8 +76,6 @@ class ListRow extends React.Component {
     var edit = "users/" + this.props.user.id + "/edit";
     var del = "users/" + this.props.user.id;
     var btnClass = "btn btn-danger";
-    console.log(this.props.user.id + "\n")
-    console.log(this.props.current)
     if (this.props.user.id == this.props.current) {
       btnClass = "btn btn-danger disabled"
     }
@@ -72,4 +92,3 @@ class ListRow extends React.Component {
 		)
 	}
 }
-
